LS: Move all formatter logic into `ide` module
commit-id:e121dc79
