+++ crates/cairo-lang-language-server/src/ide/mod.rs
@@ -1,2 +1,3 @@
 pub mod completion;
+pub mod formatter;
 pub mod semantic_highlighting;
